backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-ollama
pypdf
chromadb
umap-learn
numpy
scikit-learn
